+++ src/main/java/org/example/domain/enums/CupSize.java
@@ -1,0 +1,16 @@
+package org.example.domain.enums;
+
+public enum CupSize {
+    CHICO(1), MEDIANO(2), GRANDE(3);
+
+    private final int maxScoops;
+
+    CupSize(int maxScoops) {
+        this.maxScoops = maxScoops;
+    }
+
+    public int getMaxScoops() {
+        return maxScoops;
+    }
+}
+