+++ src/main/java/org/example/domain/enums/Flavor.java
@@ -1,0 +1,11 @@
+package org.example.domain.enums;
+
+public enum Flavor {
+    VAINILLA,
+    CHOCOLATE,
+    FRUTILLA,
+    MENTA_GRANIZADA,
+    GRANIZADO,
+    DULCE_DE_LECHE
+}
+