+++ src/main/java/org/example/domain/enums/Topping.java
@@ -1,0 +1,9 @@
+package org.example.domain.enums;
+
+public enum Topping {
+    CHIPS_CHOCOLATE,
+    SALSA_CARAMELO,
+    SALSA_FRUTILLA,
+    NUECES
+}
+