+++ src/main/java/org/example/domain/interfaces/Validatable.java
@@ -1,4 +1,4 @@
-package org.example.domain;
+package org.example.domain.interfaces;
 
 public interface Validatable {
     void validate() throws org.example.exceptions.ValidationException;